Move socket_util module into compat
